# proxy: add request and response size metrics