--- a/metrics/all_kind.go
+++ b/metrics/all_kind.go
@@ -64,6 +64,11 @@ func (a *All) MeasureAllFiltersRequest(routeId string, start time.Time) {
 	a.codaHale.MeasureAllFiltersRequest(routeId, start)
 }
 
+func (a *All) MeasureBackendRequestHeader(host string, size int) {
+	a.prometheus.MeasureBackendRequestHeader(host, size)
+	a.codaHale.MeasureBackendRequestHeader(host, size)
+}
+
 func (a *All) MeasureBackend(routeId string, start time.Time) {
 	a.prometheus.MeasureBackend(routeId, start)
 	a.codaHale.MeasureBackend(routeId, start)
@@ -87,6 +92,11 @@ func (a *All) MeasureAllFiltersResponse(routeId string, start time.Time) {
 func (a *All) MeasureResponse(code int, method string, routeId string, start time.Time) {
 	a.prometheus.MeasureResponse(code, method, routeId, start)
 	a.codaHale.MeasureResponse(code, method, routeId, start)
+}
+
+func (a *All) MeasureResponseSize(host string, size int64) {
+	a.prometheus.MeasureResponseSize(host, size)
+	a.codaHale.MeasureResponseSize(host, size)
 }
 
 func (a *All) MeasureProxy(requestDuration, responseDuration time.Duration) {