--- a/proxy/metrics_test.go
+++ b/proxy/metrics_test.go
@@ -2,7 +2,11 @@ package proxy_test
 
 import (
 	"fmt"
+	"net"
 	"net/http"
+	"net/http/httptest"
+	"strconv"
+	"strings"
 	"testing"
 	"time"
 
@@ -11,6 +15,7 @@ import (
 	"github.com/zalando/skipper/eskip"
 	"github.com/zalando/skipper/filters/builtin"
 	"github.com/zalando/skipper/metrics/metricstest"
+	"github.com/zalando/skipper/net/dnstest"
 	"github.com/zalando/skipper/proxy"
 	"github.com/zalando/skipper/proxy/proxytest"
 	"github.com/zalando/skipper/routing"
@@ -93,5 +98,98 @@ func TestMeasureProxyWatch(t *testing.T) {
 		assert.InDelta(t, measures["proxy.total.duration"][0].Seconds(), 0.001, 0.001)
 		assert.InDelta(t, measures["proxy.request.duration"][0].Seconds(), 0.001, 0.001)
 		assert.InDelta(t, measures["proxy.response.duration"][0].Seconds(), 0.001, 0.001)
+	})
+}
+
+func TestMeasureResponseSize(t *testing.T) {
+	dnstest.LoopbackNames(t, "foo.skipper.test", "bar.skipper.test")
+
+	m := &metricstest.MockMetrics{}
+	defer m.Close()
+
+	backend := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		size, err := strconv.Atoi(r.URL.Query().Get("size"))
+		if err != nil {
+			http.Error(w, "Invalid size parameter", http.StatusBadRequest)
+			return
+		}
+		w.Write([]byte(strings.Repeat("x", size)))
+	}))
+	defer backend.Close()
+
+	p := proxytest.Config{
+		Routes: eskip.MustParse(fmt.Sprintf(`
+			foo: Host("^foo[.]skipper[.]test") -> "%s";
+			bar: Host("^bar[.]skipper[.]test") -> "%s";
+		`, backend.URL, backend.URL)),
+		ProxyParams: proxy.Params{Metrics: m},
+	}.Create()
+	defer p.Close()
+
+	client := p.Client()
+	get := func(url string) {
+		t.Helper()
+		rsp, _, err := client.GetBody(url)
+		require.NoError(t, err)
+		require.Equal(t, 200, rsp.StatusCode)
+	}
+
+	fooHost := net.JoinHostPort("foo.skipper.test", p.Port)
+	barHost := net.JoinHostPort("bar.skipper.test", p.Port)
+
+	get("http://" + fooHost + "/?size=1000")
+	get("http://" + fooHost + "/?size=1234")
+	get("http://" + barHost + "/?size=555")
+	get("http://" + barHost + "/?size=77777")
+
+	m.WithValues(func(values map[string][]float64) {
+		assert.Equal(t, []float64{1000, 1234}, values[fmt.Sprintf("response.%s.size_bytes", fooHost)])
+		assert.Equal(t, []float64{555, 77777}, values[fmt.Sprintf("response.%s.size_bytes", barHost)])
+	})
+}
+
+func TestMeasureBackendRequestHeader(t *testing.T) {
+	dnstest.LoopbackNames(t, "foo.skipper.test", "bar.skipper.test")
+
+	m := &metricstest.MockMetrics{}
+	defer m.Close()
+
+	backend := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
+	defer backend.Close()
+
+	var (
+		fooHeader = strings.Repeat("A", 100)
+		barHeader = strings.Repeat("B", 200)
+	)
+
+	p := proxytest.Config{
+		Routes: eskip.MustParse(fmt.Sprintf(`
+			foo: Host("^foo[.]skipper[.]test") -> setRequestHeader("Foo", "%s") -> "%s";
+			bar: Host("^bar[.]skipper[.]test") -> setRequestHeader("Bar", "%s") -> "%s";
+		`, fooHeader, backend.URL, barHeader, backend.URL)),
+		RoutingOptions: routing.Options{FilterRegistry: builtin.MakeRegistry()},
+		ProxyParams:    proxy.Params{Metrics: m},
+	}.Create()
+	defer p.Close()
+
+	client := p.Client()
+	get := func(url string) {
+		t.Helper()
+		rsp, _, err := client.GetBody(url)
+		require.NoError(t, err)
+		require.Equal(t, 200, rsp.StatusCode)
+	}
+
+	fooHost := net.JoinHostPort("foo.skipper.test", p.Port)
+	barHost := net.JoinHostPort("bar.skipper.test", p.Port)
+
+	get("http://" + fooHost)
+	get("http://" + barHost)
+
+	m.WithValues(func(values map[string][]float64) {
+		fooSize := values[fmt.Sprintf("backend.%s.request_header_bytes", fooHost)][0]
+		barSize := values[fmt.Sprintf("backend.%s.request_header_bytes", barHost)][0]
+
+		assert.Equal(t, barSize-fooSize, float64(len(barHeader)-len(fooHeader)))
 	})
 }