Measure backend request header and response sizes.

Request header size calculation benchmark results
(compared to measurement of written bytes):
```
$ go test ./net/ -run=none -bench=BenchmarkSizeOfRequestHeader -count=10 | benchstat -col .fullname -row=name -
goos: linux
goarch: amd64
pkg: github.com/zalando/skipper/net
 │ SizeOfRequestHeader/exact-8 │      SizeOfRequestHeader/fast-8      │
 │           sec/op            │    sec/op     vs base                │
                14752.5n ± 14%   720.2n ± 17%  -95.12% (p=0.000 n=10)

 │ SizeOfRequestHeader/exact-8 │     SizeOfRequestHeader/fast-8     │
 │            B/op             │    B/op     vs base                │
                   8863.0 ± 0%   704.0 ± 0%  -92.06% (p=0.000 n=10)

 │ SizeOfRequestHeader/exact-8 │     SizeOfRequestHeader/fast-8     │
 │          allocs/op          │ allocs/op   vs base                │
                   16.000 ± 0%   1.000 ± 0%  -93.75% (p=0.000 n=10)
```

Metrics collection is only implemented for Prometheus, see https://github.com/zalando/skipper/issues/3530

Also rename some existing prometheus histograms for consistency.